backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
pymssql==2.2.11
pandas==2.2.0
openpyxl==3.1.2
python-multipart==0.0.6
pydantic==2.5.3
# pyodbc (opcional, requerido solo para Windows Authentication en SQL Server)
# pip install pyodbc

# Dependencias para envío de correos y Google Drive (opcionales)
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
